base_batch_predictions_iris_sample/requirements.txt
google-cloud-sql
google-auth
google-cloud-bigquery
google-cloud-aiplatform
psycopg2